client creation as util
