+++ internal/integ/cluster/setup.go
@@ -7,8 +7,8 @@ import (
 	"log/slog"
 	"os"
 
-	"github.com/exoscale/exoscale/csi-driver/internal/integ/client"
 	"github.com/exoscale/exoscale/csi-driver/internal/integ/flags"
+	"github.com/exoscale/exoscale/csi-driver/internal/integ/util"
 
 	exov3 "github.com/exoscale/egoscale/v3"
 )
@@ -102,7 +102,7 @@ func exitApplication(msg string, err error) {
 }
 
 func ConfigureCluster(ctx context.Context, createCluster bool, name, zone string) (*Cluster, error) {
-	client, err := client.CreateEgoscaleClient()
+	client, err := util.CreateEgoscaleClient()
 	if err != nil {
 		return nil, fmt.Errorf("error creating egoscale v3 client: %w", err)
 	}
